+++ src/SelectedStateParks/SelectedStateParks.js
@@ -1,12 +1,33 @@
 import React from "react";
 import PropTypes from "prop-types";
 import Card from "../Card/Card";
+import axios from "axios";
+import { useNavigate } from "react-router-dom";
 import "./SelectedStateParks.css";
 
-function SelectedStateParks({ parks, selectedState, fetchParkDetails }) {
+function SelectedStateParks({ showResults, parks, selectedState, apiLink }) {
   if (!parks || parks.length === 0) {
-    return <div>No parks available.</div>;
+    return <h3>No parks available.</h3>;
   }
+
+  const navigate = useNavigate();
+  const fetchParkDetails = async (park_code) => {
+    try {
+      const response = await axios.get(`${apiLink}/${park_code}`);
+      const selectedPark = response.data.data;
+
+      if (selectedPark) {
+        setParkCode(selectedPark.attributes.park_code);
+        setFoundPark(selectedPark);
+      } else {
+        console.error(`Park with id ${park_code} not found.`);
+      }
+      navigate(`/Parks/${selectedState}/${park_code}`);
+    } catch (error) {
+      console.error("Error fetching park details:", error);
+    } finally {
+    }
+  };
 
   const filteredState = parks.filter(
     (results) => results.attributes.states === selectedState
